+++ CHANGELOG.md
@@ -10,6 +10,7 @@ All notable changes to this project will be documented in this file. This projec
 
 -   Focus input field when showing the main window.
 -   Disable player buttons if no text to speak.
+-   Always speak translation on its language.
 
 ## [2.2.0](https://github.com/crow-translate/crow-translate/tree/2.2.0) (2019-08-31)